+++ src/main/java/com/mysite/sbb/answer/Answer.java
@@ -1,12 +1,12 @@
-package com.mysite.sbb;
+package com.mysite.sbb.answer;
 
+import com.mysite.sbb.question.Question;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 import static jakarta.persistence.GenerationType.*;
 
@@ -14,20 +14,19 @@ import static jakarta.persistence.GenerationType.*;
 @Setter
 @Entity
 @NoArgsConstructor
-public class Question {
+public class Answer {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Integer id;
-
-    @Column(length = 200)
-    private String subject;
 
     @Column(columnDefinition = "TEXT")
     private String content;
 
     private LocalDateTime createDate;
 
-    @OneToMany(mappedBy = "question",cascade = CascadeType.REMOVE)
-    private List<Answer> answerList;
+
+    @ManyToOne
+    private Question question;
+
 }
