+++ src/main/java/com/mysite/sbb/question/Question.java
@@ -1,11 +1,14 @@
-package com.mysite.sbb;
+package com.mysite.sbb.question;
 
+import com.mysite.sbb.answer.Answer;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 import static jakarta.persistence.GenerationType.*;
 
@@ -13,19 +16,20 @@ import static jakarta.persistence.GenerationType.*;
 @Setter
 @Entity
 @NoArgsConstructor
-public class Answer {
+public class Question {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Integer id;
+
+    @Column(length = 200)
+    private String subject;
 
     @Column(columnDefinition = "TEXT")
     private String content;
 
     private LocalDateTime createDate;
 
-
-    @ManyToOne
-    private Question question;
-
+    @OneToMany(mappedBy = "question",cascade = CascadeType.REMOVE)
+    private List<Answer> answerList = new ArrayList<>();
 }
