+++ src/test/java/com/mysite/sbb/SbbApplicationTest.java
@@ -1,5 +1,9 @@
 package com.mysite.sbb;
 
+import com.mysite.sbb.answer.Answer;
+import com.mysite.sbb.answer.AnswerRepository;
+import com.mysite.sbb.question.Question;
+import com.mysite.sbb.question.QuestionRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -15,6 +19,8 @@ class SbbApplicationTest {
 
     @Autowired
     private QuestionRepository questionRepository;
+    @Autowired
+    private AnswerRepository answerRepository;
 
     @Test
     void testJpa() {
@@ -33,11 +39,19 @@ class SbbApplicationTest {
         List<Question> all = questionRepository.findAll();
         assertThat(all.size()).isEqualTo(2);
 
-        Optional<Question> byId = questionRepository.findById(1);
+        Optional<Question> byId = questionRepository.findById(2);
         if (byId.isPresent()) {
             Question q = byId.get();
-            assertThat(q.getSubject()).isEqualTo("sbb가 무엇인가요?");
+            assertThat(q.getSubject()).isEqualTo("스프링부트 모델 질문입니다.");
+
+            Answer a = new Answer();
+            a.setContent("네 자동으로 생성됩니다.");
+            a.setQuestion(q);
+            a.setCreateDate(LocalDateTime.now());
+            answerRepository.save(a);
+
         }
+
 
         Question findQuestion = questionRepository.findBySubject("sbb가 무엇인가요?");
         assertThat(findQuestion.getId()).isEqualTo(1);
